finished preparation for supporting CSV
