+++ include/BVDSettings.h
@@ -10,13 +10,23 @@
 class INIReader;
 struct BVDContainmentStrategy;
 
-
-
+enum fileMode{
+	single_file = 0,
+	multi_file = 1
+};
+struct OutputSettings{
+    bool overwrite;
+    std::string fileExtension;
+    std::string fullFilePath;
+    std::string fileprefix;
+    std::string path;
+    fileMode mode;
+};
 class BVDSettings{
 public:
-    static BVDSettings* sharedInstance(INIReader* reader);
+    static BVDSettings* sharedInstance(INIReader* reader = nullptr);
     StrategyQueue strategies;
-
+    OutputSettings outputSettings;
 private:
     INIReader* reader;
     BVDSettings(INIReader* reader);
@@ -36,15 +46,20 @@ private:
 
 
     void initializeStrategies();
+
     void initializeStrategy(const std::string& name);
 
-
+    void initializeOutputSettings();
+    static fileMode iniInputToFileMode(std::string& fileMode,bool shallExit=true);
     class CGuard
 	{
 	public:
 	   ~CGuard();
 
 	};
+
 };
+
+
 
 #endif
