+++ iniFiles/pascalTest.ini
@@ -164,7 +164,7 @@ filetype=hdf5
 fileprefix=testing
 #mode=single_file
 mode=split_file
-path=../../Tests/
+path=../
 overwrite=true
 
 [rng]
