+++ src/FileHandler.cpp
@@ -1,12 +1,15 @@
 #include "FileHandler.h"
 #include "projectImports/inih/cpp/INIReader.h"
-FileHandler::FileHandler(INIReader& reader){
-		this->overwrite = reader.GetBoolean("output", "overwrite", false);
-		this->fileprefix = reader.Get("output", "fileprefix", "output_");
-		this->path = reader.Get("output", "path", "./");
-		this->setPath();
-	}
-FileHandler::~FileHandler(){} 
+#include "BVDSettings.h"
+FileHandler::FileHandler(){
+	this->overwrite = BVDSettings::sharedInstance()->outputSettings.overwrite;
+    this->fileprefix = BVDSettings::sharedInstance()->outputSettings.fileprefix;
+    this->path = BVDSettings::sharedInstance()->outputSettings.path;
+
+    this->fullFilePath = BVDSettings::sharedInstance()->outputSettings.fullFilePath;
+	this->setPath();
+}
+FileHandler::~FileHandler(){}
 
 void FileHandler::setPath(){
 	this->fullFilePath = this->path + this->fileprefix + this->fileExtension;
